Add files via upload
This button manager keeps track of which buttons are highlighted and calls the appropriate unlock function for the highlighted button. It has been updated so that the unlock button follows a different coloring logic to the other buttons and also to disallow the user from calling unlock when no button has been selected.